rename mf pack
